refactor: canvas models to feature-based architecture
- Move canvas_color.dart to features/canvas/models/
- Move tool_mode.dart to features/canvas/models/
- Update all import paths in affected files
- Begin transition from layer-based to feature-based structure"
