+++ lib/features/canvas/models/tool_mode.dart
