+++ lib/widgets/canvas/canvas_toolbar.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:scribble/scribble.dart';
 
-import '../../models/canvas_color.dart';
+import '../../features/canvas/models/canvas_color.dart';
+import '../../features/canvas/models/tool_mode.dart';
 import '../../models/custom_scribble_notifier.dart';
-import '../../models/tool_mode.dart';
 import 'color_button.dart';
 import 'drawing_mode_toolbar.dart';
